+++ .gitignore
@@ -49,6 +49,7 @@ build_config.yaml
 # Tests
 .coverage
 coverage-reports
+merge_master_test
 
 # Doc pdf file
 antaresxpansionuserguide.pdf
